pyproject.toml
[project]
name = "simplenet_learner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "INAGAKI Tatsuya", email = "inagaki.tatsuya@jp.panasonic.com" }
]
requires-python = ">=3.11"
dependencies = [
    "colorlog>=6.9.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.5.0.post0",
    "rich>=13.9.4",
    "scipy>=1.15.1",
    "torch>=2.4.1",
    "torchinfo>=1.8.0",
    "torchmetrics>=1.6.0",
    "torchvision>=0.19.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "simplenet-learner",
]

[project.scripts]
simplenet_learner = "simplenet_learner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
simplenet-learner = { workspace = true }
